+++ assets/js/src/newsletters/automatic_emails/listings.jsx
@@ -2,11 +2,10 @@ import React from 'react';
 import ReactStringReplace from 'react-string-replace';
 
 import Listing from 'listing/listing.jsx';
-import confirmAlert from 'common/confirm_alert.jsx';
 import Tags from 'common/tag/tags';
 import Toggle from 'common/form/toggle/toggle';
 import { ScheduledIcon } from 'common/listings/newsletter_status';
-import { checkMailerStatus, addStatsCTAAction } from 'newsletters/listings/utils.jsx';
+import { checkMailerStatus, addStatsCTAAction, confirmEdit } from 'newsletters/listings/utils.jsx';
 import Statistics from 'newsletters/listings/statistics.jsx';
 import NewsletterTypes from 'newsletters/types';
 import classNames from 'classnames';
@@ -100,21 +99,6 @@ const bulkActions = [
     onSuccess: messages.onTrash,
   },
 ];
-
-const confirmEdit = (newsletter) => {
-  const redirectToEditing = () => {
-    window.location.href = `?page=mailpoet-newsletter-editor&id=${newsletter.id}`;
-  };
-
-  if (newsletter.type === 'automatic' && newsletter.status === 'active') {
-    confirmAlert({
-      message: MailPoet.I18n.t('confirmAutomaticNewsletterEdit'),
-      onConfirm: redirectToEditing,
-    });
-  } else {
-    redirectToEditing();
-  }
-};
 
 let newsletterActions = [
   {
